Backend_Main_SpringBoot. Corregidas tablas de migración relacionadas con Pistas para cambiar id por bigint

+++ backend_main_springboot/src/main/resources/db/migration/V3__create_core_feature_tables.sql
@@ -20,7 +20,7 @@ CREATE TABLE suscripciones (
 );
 
 CREATE TABLE pistas (
-    id int PRIMARY KEY AUTO_INCREMENT,
+    id bigint PRIMARY KEY AUTO_INCREMENT,
     nombre varchar(255),
     slug varchar(255) UNIQUE,
     descripcion text,
